refactor: move UserWithAccessToken type to src/system
