+++ src/system/index.ts
@@ -1,0 +1,3 @@
+import type { User } from "next-auth"
+
+export type UserWithAccessToken = User & { accessToken: string }
